Allow any uid/gid to be set for vertica pods (#556)
This changes the v2 image, which is used by vclusterops deployments, to
allow it to be run with any UID/GID. The IDs are picked at runtime.
Previous to this the image would hard code the IDs to 5000. If you
wanted to run with a different ID you had to recreate the image. Now you
can change the IDs at runtime by filling in the
`spec.podSecurityContext`. For instance, if you want to use 3500 as the
IDs, you fill it in like this:

```
apiVersion: vertica.com/v1
kind: VerticaDB
metadata:
  name: v
spec:
  image: vertica/vertica-k8s:24.1.0-0
  communal:
    path: "s3://vertica-fleeting/mspilchen"
  podSecurityContext:
    runAsUser: 3500
    runAsGroup: 3500
  subclusters:
    - name: pri
      size: 3
```      
      
A couple of notes about this:
- it only works with vclusterops deployments. The old admintools
deployment still rely on the hard-coded 5000 value.
- once the IDs are picked we have a webhook that prevents the value from
changing. We cannot allow them to change because you may end up with
permission problems.
- if the podSecurityContext is omitted the operator will pick sane
values at runtime. For non-OpenShift it will pick 5000. In OpenShift it
will read the namespace metadata to know the range of IDs to use.
- we dropped the custom SCC that we had for OpenShift. Our vertica
deployment can run with anyuid scc that comes with OpenShift. We have
plans to run in restricted-v2 scc but we need to remove the s6-overlay
init program as that needs elevated privileges to run.
- the operator will create temporary files. These are now written to
/tmp because /home/dbadmin in the pod may not be writable.