Re-enable restart-node-multi-sc e2e test (#562)
`restart-node-multi-sc` is re-enabled after the `AddNodeLicenseFail` can
be recognized in `vclusterops` deployment just like in `admintools`
deployment. The step 21 "check log rotation" is deleted because it uses
`admintools` and the server logrotate functionality is currently being
re-implemented, we need to have a new dedicated test for log rotation
after the new implementation of logrotate is completed.