Run the NMA as a longrun command in the v2 image (#569)
The key change here is that we now start the NMA synchronously in the v2
image, instead of asynchronously. This means that the s6 init process
will monitor the result and restart the NMA if it stops. This is
necessary for the istio testcase because network access may not be
available when the istio sidecar starts, which could cause the NMA to
fail. By having s6 restart the NMA, we can make that testcase more
resilient.

A few other changes are included: 
- upgrading kind to latest version (0.20.0)
- changes to the istio testcase so that it runs with and without the
webhook. When the webhook is running we had to ensure the port does not
allow mTLS. All access to the webhook comes from the k8s api-server,
which is always outside of the mesh. So, it's untrusted (from istio's
point of view) and can't use mTLS.