Handle missing versions as an error (#602)
The controller will obtain the Vertica version and save it as an
annotation in the VerticaDB object. This is expected to be completed
before the create db operation. This will allow us to take the
appropriate actions based on the Vertica version we are using. However,
this is not always reliable due to the caching of the Kubernetes client.
Previously, we tolerated missing versions, but I am now changing some of
these to errors in cases where we need the version.

This was causing an issue when enabling spread encryption during create
db. In older Vertica versions, or if the version is missing, the server
would restart. By handling missing versions as an error, we will take
the correct action. I believe this will also improve the reliability of
the e2e tests.

I am also removing some temporary code that included the Go routine ID
with each reconcile. I am generating a UUID and using that instead. This
is a more permanent solution.