Use k8s.io package to manage VerticaDB status condition (#600)
In the v1beta1 API we added our own status condtions to the VerticaDB.
Now in v1, we want to use the Condition struct from
`k8s.io/apimachinery/pkg/apis/meta/v1`. This PR is to remove
v1.VerticaDBCondition and all related code in favor of
`metav1.Condition`. It will make it easier to manage status condition
with provided methods