Run scrutinize for vclusterops deployments (#605)
For vclusterops deployments, the server team developed a new method for
scrutinize using vcluster. This uses the new scrutinize method for those
deployments. We had a collection script that we run at the end of e2e
tests. This has been updated. We also call this script in a few
testcases to verify that scrutinize runs correctly.