Intermittent failure in verify-server-logrotate e2e test (#610)
An intermittent failure in the verify-server-logrotate e2e test occurs
because I forgot to clean up the EnableLogRotate parameter. Sometimes,
the time issue occurs when we switch from the off to on time service. We
need to include a step to clean up the EnableLogRotate parameter. After
this change, the test will be stable.