Fix intermittent e2e test failure (#612)
This is another attempt at fixing the timing scenario in
revive-with-different-local-paths. I suspect that the assertion for step
50 has a small timing window -- we need to provision the pods and still
see the paths as invalid in the VerticaDB, but the operator can update
the paths pretty quick. I am just going to check for the paths and not
bother with verifying the pods any more.