capture-scrutinize.sh to use IP instead of host
We have a helper script called capture-scrutinize.sh that will gather
scrutinize during e2e runs. A few e2e tests use it and require that
scrutinize be successful. One of those testcases, nma-cert-mounts, has
failed because it couldn't talk to the NMA. It was running this right
after a restart, but it was sending the request to an old IP. We pass in
the host name to scrutinize. I suspect it resloved the host using a
stale DNS cache and got the old IP. I am changing the script to use IPs
instead of hosts.