Fix intermittent auto-restart-vertica test failure (#613)
That test will create an EventTrigger that will run when the
AutoRestartVertica status condition goes to False. We change the status
condition, but we don't always see the EventTrigger creating the job.
Each time it fails, looking at the log, you can see that although
AutoRestartVertica is False(checked with kubectl), the value used for
comparison by eventtrigger is true: `{"et":
"kuttl-test-novel-toad/hello-world", "expected": "False", "found":
"True", "refObjectName": "v-auto-restart-vertica", "condition":
"AutoRestartVertica"}`, so we are somehow picking the wrong value. I
suspect it is because we are using a `v1beta1` vdb object to get the
status from k8s so instead I used a `v1`.