Fix error communal storage path is invalid (#617)
We include the UID in the communal path to generate a unique path for
each new instance of vdb. To do that, we append the UID to the path this
way `path/uid`. The problem is if the path ends with `/` like
`s3://vertica-fleeting/k8s/`, we end up with 2 slashes like
`s3://vertica-fleeting/k8s//uid` which will fail vclusterops communal
storage path check with error: `communal storage path is invalid: use an
absolute local path or a correct remote url path`.
So I removed the trailing slash from the path before appending the UID.