Allow NMA certs to be read from GSM (#618)
The key change here is to allow the NMA certs to be read in from Google
Secret Manager (GSM). This is triggered based on a secret path reference
given for the `spec.nmaTLSSecret` secret. The secret path reference for
GSM is `gsm://`. So, the secret name has to be filled in like
`gsm://projects/1680942424/secrets/mspilchen/versions/6`.

Some other changes that were made:
- secrets stored in GSM need to be saved in a JSON format. The values
for each key in the JSON doc is expected to be a byte-array so must be
stored as base64 encoded. Previously the values were stored as plaintext
strings. Using base64 encoded will allow us to easily store multi-line
certs in them.
- generalized the secret reader in the operator so that it can use GSM
or K8s depending on the path reference. This was added in a new struct
called `MultiSourceSecretFetcher` in the `cloud` package.
- new helm chart parameters to allow you to specify annotations for the
service account used to deploy the operator. This was added as a
convenience because the annotation is necessary to link a GCP service
account with the k8s service account.