Speed up revive prepass (#619)
During the revive prepass, the operator will compare the paths in the
VerticaDB with the data in communal storage. If there is a discrepancy,
the operator will patch the VerticaDB so that the local paths match what
is expected. This is done instead of failing the revive.

Once the StatefulSet is updated, it is then updated in the pod. Prior to
this change, pods that were on the old revision would be deleted.
However, this could be a slow process because it was possible for pods
to come back with the old revision, depending on which pods had the
newest revision.

To avoid this, the StatefulSet is simply deleted in this change. This is
a quicker process.

Closes #568