Allow custom superuser name for admintools deployments (#622)
This allows a custom superuser name to be used. It builds on the work done in
#578. That issue was specifically for vclusterops deployments. This one is
specific to admintools deployments. It fixes a few code paths that were
specific to admintools deployments. Note: the admintools image hard codes the
superuser name as dbadmin, so you will still need a custom image to run with
different named superuser. This change is strictly for the operator.