Fix e2e test failures (#628)
This tries to fix a small timing window in
revive-with-different-local-paths. There is a small window, 13 seconds
in the failed case, when the k8s objects are in the correct state to
match step 50 assert. If we transition too fast, then we don't catch the
assertion. I am trying to fix this by reducing what we are checking in
the assertion.