Fix restart reconciler requeue logic for vclusterops
This is to fix intermittent failures in the pending-pod e2e testcase,
when run in vclusterops mode. The restart reconciler needs to requeue if
pods that it needs aren't running. We were too aggressive with
vclusterops because we triggered on installed state, which may or may
set correctly. This changes the check for vclusterops and only looks at
dbExists state. It also correctly sets the installed state for
vclusterops in all cases.