Online upgrade to account for terminating pods (#633)
The controller was not handling terminating pods correctly. This could
lead to problems during online upgrades, such as not properly restarting
primaries before taking down secondaries.

The issue occurs when the operator deletes pods with the old image. If
the pod is still running immediately after the deletion, the operator
may skip the restart and move on to the next phase of the upgrade. This
is a small window because the termination grace period is 0 seconds, but
we still see it occasionally in the e2e runs.

To test this issue, I added a new annotation
(vertica.com/termination-grace-period-seconds) that makes it easier to
hit this timing scenario. This annotation causes the pod to be in the
termination step for a longer period of time.

The solution was to check for pod termination in podfacts and treat the
pod as not running if it is set.