Allow upgrade through a deployment change
If upgrading from an admintools deployment to a vclusterops deployment,
this change is needed if doing that via an online upgrade. We can't
handle an online upgrade in this fashion because the NMA isn't running
in pods still doing an admintools style of deployment. We have to take
down any pods running admintools deployments. This will make the online
upgrade offline, but at leaste then we can proceed.