Generate prometheus_metrics_latest.json (#645)
This adds a file to the repo called prometheus_metrics_latest.json. It
is a summary of all of the metrics exposed by the operator. The purpose
of this file is for doc writers to use it to include details about the
metrics found in the operator.

There is an e2e test (metrics-no-auth) that will ensure this file stays
up to date.

@rs-vertica I included some metrics in the json that the operator
inherits from the SDK. Here is some more detail about those metrics
here: https://book.kubebuilder.io/reference/metrics-reference. It looks
like it is just a summary of the help text. I can omit those and just
include the ones we add. We don't have any control over the help text.
So, if the ones we inherit have any spelling and/or grammatical errors
then we would need to manually fix them. Let me know if you want me to
omit some.