Fixes for flaky tests (#649)
After monitoring the test results over the last week, I noticed two
things that I am fixing:
- the e2e test crash-before-createdb, is suppose to kill pods in the
middle of a create db. But sometimes the database is created before the
kills if it is really quick. I am improving the testcase so that it can
tolerate this.
- unit test for secrets package can sometimes fail because a namespace
it depends on doesn't exist. I am changing the testcase so that it
explicitly creates the namespace it needs.