VCluster API refresh: allow create db to proceed if db is already running (#652)
This builds on a change made in vclusterops. It can now emit a specific
error if it discovers that the database is already running. We add a
feature flag to tolerate this (the default is true). This can allow us
to proceed if we are in an environment with few resources available,
which can cause instability during the creation process. If we have
previously progressed far enough that Vertica is running, we will skip
the create db step and proceed to the next reconciler. This will update
the status condition to indicate that the database has been initialized.