Include omitted release manifests (#660)
The build release assets workflow was not including all of the manifests
that we wanted to include with the operator GitHub release. This has
been corrected. Additionally, the operator.yaml that we generate for the
release was not using the correct image name for our public operator
image. This has also been fixed. These issues were discovered while
releasing the 2.0.0 of the operator.