Support for running the NMA as a sidecar (#654)
This adds initial support in the operator to be able to run the NMA as a
sidecar. This is the default way of deploying the statefulset with the
v1 VerticaDB API. This requires a server version of at least 24.2.0.
When running 24.1.0 you must set this annotation in the VerticaDB:
vertica.com/run-nma-in-sidecar: "false". In 24.2.0 or newer, this
annotation can be omitted.

We now properly setup the statefulset for the sidecar deployment. The
same image is used for both the NMA and server containers. The NMA
container is the first container in the pod spec template. This was done
because we want it to be the default one when entering via 'kubectl
exec'. This is important when collecting scrutinize as that must be run
in the NMA container.

It originally had some container changes, like removing the s6 init
process, but we had to back that out as we need to continue to support
running the monolithic container in 24.2.0 for the time being.

We added a temporary hack to account for the vertica process turning
into a daemon. We add a sleep infinity after running the vertica start
command. This ensures the pod never exits under normal circumstances.
This will be removed in a subsequent PR once the server side changes are
fixed.

---------

Co-authored-by: Matt Spilchen <mspilchen@opentext.com>