Fixes for OLM style deployments (#662)
When publishing the 2.0.0 operator in operatorhub.io/OpenShift, I came
across a few issues that were preventing it. This change addresses that.
A summary of the changes are:
- use a new channel for the 2.0.0 release. We are switching from a
namespace scoped operator to a cluster scoped operator, so we need to
define a new channel to prevent upgrading from the 1.x.x operator to the
2.x.x operator. The new channel is called v2-stable (the old channel was
stable).
- change the minimum Kubernetes version from 1.21.1 to 1.21.0. There is
no change in our support, but to get the operator on OpenShift we had to
drop the x.x.1 to be compatible with OCP 4.8.
- don't allow the operator CRD write permissions of OLM style
deployments. We added CRD write permissions so that the operator can
update the conversion webhook, which is stored in the CRD, with the CA
bundle of the webhook cert. OLM style deployments handle this
themselves, so the operator does not need to do anything. This also
required us to upgrade OLM in our tests from 0.22.0 to 0.26.0. The old
version we were using didn't automatically update the conversion
webhook.