Add support for rockylinux 8 in our server container (#665)
This adds support for rockylinux 8 in our container builds. Now we can
switch between version 8 and 9 through make variables. Rockylinux 8 is
the default for new server containers going forward. This applies both
to the v1 and v2 vertica images. The first publicly available image with
rockylinux 8 will be 24.2.0.

In order to support both versions, the udx e2e testcase setup had to
change a bit. I think its easier to run that now on your local machine.

vdb-gen had to compile in static libraries for libc so that one compiled
on a newer OS can run on rockylinux 8.