Fix timing with NMA sidecar and drop node (#666)
This change addresses a timing issue with the new NMA sidecar and drop
node. If the drop node process is interrupted, it is possible that the
next reconcile iteration will attempt to restart the pod. However, this
restart will fail because the pod may no longer be in the database,
which will cause the restart to fail. This change addresses this issue
by skipping pods in this state during restart.