Don't use NMA sidecar in our CI when running 24.1.0 (#668)
When our server builds a new hotfix for 24.1.0 it will kick off a CI run
and try to run some of the tests with the NMA sidecar. The NMA sidecar
is only supported in 24.2.0 and above. This changes disables NMA sidecar
when it detects you are running with 24.1.0.

This also cleans up some logic related to guessing the base image for a
version. We had a complicated set of instructions for picking the right
base image so that our upgrade tests can adhere to the upgrade path. The
upgrade path logic has since been removed, which makes it easier to pick
the base image. So, that has been cleaned up.