Remove 'sleep infinity' hack for NMA sidecar support (#667)
When we did the first drop of the NMA sidecar support (#654), we had to
added a 'sleep infinity' to the vertica command because that process
would always daemonize itself. With a recent server change that prevents
vertica from becoming a daemon, we can now remove the hack.

This also adds health probes to the NMA container by checking the
/v1/health endpoint.

A couple of other minor changes went into this PR:
- a new annotation was added (`vertica.com/skip-deployment-check`) to
avoid checking if the image was built for admintools or vclusterops
deployments. This is useful if using a build that can do either types of
deployments.
- set more environment variables in the NMA container that were set in
the server. The main driver for this is it allows you to run vsql
without having to specify the username.