CI work for expanded NMA sidecar container test coverage (#672)
This CI work is part of the NMA sidecar effort. A few test cases were
sensitive to the containers present, so they've been added to a new leg
(e2e-leg-7) that only runs on server releases 24.2.0 or later.

Additionally, a new leg has been added to run against the 24.1.0
release. This is the only release that deploys with vclusterops and runs
as a monolithic container. Therefore, it was crucial to prevent any
regressions moving forward.

The NMA_RUNNING_MODE is now determined by the deployment type and
version. The hardcoding of the running mode has been removed from all
e2e legs. This will help us expand the test coverage for the NMA
sidecar.