Fix offline-upgrade-ks-0 e2e test (#681)
Small fix to a recent change I made to this test for the NMA sidecar
feature. I need to ensure we wait appropriately for the operator to
acknowledge an upgrade before moving onto the next step. We use to
verify this by checking the image, but that was removed because there is
another container to check now.