Assign resources to the NMA sidecar container (#674)
The NMA sidecar container now has default resources assigned to it in
the pod spec. Some default values have been selected: 1CPU and 250MB of
memory. These are for requests only. I intentionally left off the limit
because of the burstable nature of the NMA container.

The resources are only set if the subcluster has defined resources. This
ensures that the container can be run in low-resource environments (for
testing purposes).

New annotations have been added to facilitate resource tuning. The new
annotations are as follows:
vertica.com/nma-resources-limits-memory
vertica.com/nma-resources-limits-cpu
vertica.com/nma-resources-requests-memory
vertica.com/nma-resources-requests-cpu

Lastly, an existing e2e test case was updated. The test case was moved
to the e2e-leg-7, as it assumes the use of the NMA sidecar and can only
run on 24.2.0.