CI Updates for Node 16 deprecation (#684)
This updates actions used in the CI to new versions that aren't using
Node 16. Node 16 was deprecated in GitHub actions, so we have warnings
when we have actions using them. There are still a few remaining
deprecations, which will be fixed at a later time.

One other notable change. The verify-public-release workflow has been
updated so we have the option now to deploy the operator with kubectl.