Remove s6-overlay and NMA sidecar annotation (#689)
This removes from the v2 server container the s6-overlay init process.
Now that we have full support for running the NMA in a sidecar, this
systemd-style process isn't needed. One advantage of this change is it
allows us run in OpenShift with the restricted SCC. This makes it easier
to deploy Vertica in OpenShift because this is the default SCC. This
will be available for consumption in the v24.2.0 server to be released
in April 2024.

Also removed in this change is the `vertica.com/run-nma-in-sidecar`
annotation. This previously was needed to correctly setup the pod spec.
But the operator can automatically determine this. It defaults to using
the NMA sidecar for all vclusterOps deployments. If it finds itself on
24.1.0, then it will adjust the statefulset so that the NMA and server
are run in a single monolithic container.

This also will address different use cases as it relates to NMA
deployment and the Vertica version.
- If vclusterOps is requested on an image prior to 24.1.0, then the
container will not start. Both the NMA container, which starts
node_management_agent directly, and the server container, which starts
vertica with the `--startup-conf` setting, will fail immediately. We
cannot run anything in these pods to even get the version. We added a
new reconciler in the vdb controller that will check if the pod is in a
crash loop backoff. And if so it will log an event message under the
assumption its because we are trying to run vclusterOps on an old image.
- If vclusterOps is used with a 24.1.0 image, the server container won't
start because it is missing the `--startup-conf` setting that is new in
24.2.0. The pod facts collection will now use the NMA container if
present for these cases.