Avoid deploying the operator with default label
Any operator that is built with the operator-sdk framework will have
default selector labels added for the operator like this:
  control-plane: controller-manager

When the operator is deployed in the same namespace as other operator,
that also continue using the default label, then the service object for
the operator fails to work. It will route webhook requests to the wrong
pod.

This change is to use the following label instead:
  control-plane: verticadb-operator

I also took the opportunity to rename the operator objects. The
deployment object was renamed from verticadb-operator-controller-manager
to verticadb-operator-manager.