Ensure httpstls.json is generated for databases migrating to v24.1.0 (#698)
This fixes an issue where we didn't generate the httptls.json file for
databases migrated to 24.1.0. Without this file generated, it is
possible that the HTTPS service doesn't come up, which is needed by
vclusterOps in 24.1.0+ deployments.

We can avoid generating this file in the following circumstances:
- Databases created in 24.1.0+ will have default certs generated for
HTTPS during the bootstrap-catalog call.
- Database that have setup TLS certs in the catalog. See the docs for
that here:
https://docs.vertica.com/24.1.x/en/admin/managing-db/https-service/#certificate-authentication

The operator won't know about the second case. We have introduced a new
annotation for the first case. When
`vertica.com/https-tls-conf-generation` is set to false, we will avoid
generation. When the operator creates a database on a 24.1.0 database it
will automatically set this annotation.