Generate pod for scrutinize (#696)
This adds the logic to generate a dummy pod when a VSCR CR is created.
It does the following:
  -  checks if the VerticaDB specified in Vscr exists
- checks the server version and abort reconciliation if version is too
old
  - creates a pod

For now the pod does not run scrutinize, the logic for that will be
added on a follow-up PR. The pod composed of 2 containers, an init
container (that will collect scrutinize container later) and a main
container. Both run hard-coded image (busybox) and respectively sleep
for 5s and for infinity.
We also set some fields(initContainers, Volume,...) with values from
Vscr.