Fix for e2e-operator-upgrade tests
We changed the serviceAccount for pods with the new 2.0.2 operator. So,
some operator upgrade tests can fail due to a caching issue. The
statefulset was updated, but the sts controller was still trying to use
the old serviceaccount. I could have fixed this by increasing the
timeout of the test, but I opted to delete the sts to avoid the cache
issue.