Fixes Istio proxy sidecar initialization issue (#702)
Starting in istio 1.17, the istio proxy is now injected as the first
container. This breaks some assumptions in the operator code. It assumes
that any sidecar is injected at the end and the first two containers are
the nma and server. This changes fixes that by referencing containers by
their names rather than index. I removed all of the helpers based on
index and instead get them by name.