Fix e2e test failures (#708)
- This improves the online-upgrade-with-transient. There is a step where
it verifies that service routing during an online upgrade goes to the
correct subcluster. This can sometimes fail, even though there is
evidence the service object was updated correctly before the test ran. I
suspect there is a k8s cache issue with the service object. So, we will
retry the service object validation.
- I am trying to track down a test failure in
online-upgrade-kill-transient. We didn't restart the primary subcluster
in the order the test expects. So, I am adding debug logic to podfacts
to capture when the pod started.