Write event for admintools with a v24.2.0 server
The v24.2.0 server image, when built for vclusterops, will not work if
admintools deployment is selected. This will surface an event message if
we detect this is the case. This is important because the image will
fail to run if this is the case. This is tied into the crashloop_reconciler.