Enable the ISTIO test for vcluster deployments (#716)
I had to ensure the namespace where the operator is deployed is part of
the service mesh so that it can talk to the the REST endpoints. I also
set `holdApplicationUntilProxyStarts` to true to force the istio-proxy
container as the first container in the vertica pods. We had some bugs
around that setting that were recently fixed.