Allow namespace scoped operator (#719)
This change allows for deploying the operator at the namespace scope,
only for Helm deployments. When deployed this way, the operator will
watch only custom resources within its deployment namespace. Multiple
namespace deployments are possible.

By default, the webhook is deployed with the operator. However, with
multiple namespace scoped deployments in a cluster, the webhook can only
be deployed once.

There are two ways to handle this:
- Set `webhook.enable` to false for any subsequent operator deployment
- Deploy only the webhook (`controllers.enable` set to false) and deploy
each operator with the webhook disabled (`webhook.enable` set to false)

A new Helm chart parameter, `controllers.scope`, has been added to
control the scope of the operator, with valid values of "cluster" or
"namespace" and a default of "cluster."

The e2e-leg-2 has been converted to deploy the operator with namespace
scope.

All command-line options used by the operator have been replaced with
environment variables read from a ConfigMap. This change was made when
adding the namespace scope and controllers disable options, as this
method is more extensible.

This change removes most of the logging Helm chart parameters, leaving
only `logging.level`. Those using the logging Helm chart parameters can
set appropriate entries in the ConfigMap to retain the old behavior.

Finally, the 'make run' option has been removed. The operator can now
only be run by running a container within Kubernetes, as it was too
difficult to set up a ConfigMap and environment variables to run the
operator.