Disable operators proxy sidecar by default (#725)
The helm chart parameter `prometheus.expose` controls if the Prometheus
metrics in the operator are exposed. The old default was
`ExposeWithProxyAuth`. However, this required a
ClusterRole/ClusterRoleBinding to be created. This can be onerous,
especially in environments where creation of those is prohibited except
for cluster admins. The new default is `Disable` to make it easier to
deploy in a more restricted environment.

This also adds back logging helm chart parameters that were removed in
#719. Instead, those parameters have been deprecated and will be removed
in a future version.