CRD changes for 2.1.0 release of the operator (#728)
This merges the vnext branch into the main branch. We utilized this
branch to keep CRD changes separate until they were prepared for
release. As a result, it contains numerous modifications. The following
is a summary of the changes:
- To enable database-level restoration from a restore point, the
VerticaDB CRD has added new parameters to allow the specification of a
restore point archive for revive. For instance,
```
apiVersion: vertica.com/v1
kind: VerticaDB
metadata:
  name: verticadb-sample
spec:
  image: "opentext/vertica-k8s:24.2.0-0-minimal"
  communal:
	path: "s3://mybucket/vertica"
	endpoint: https://s3.amazonaws.com
  subclusters:
	- name: main
  	size: 3
  initPolicy: Revive
  restorePoint:
	archive: backup
	index: 1	 
```
- VerticaRestorePointsQuery is a new CRD that enables querying communal
storage to view existing restore points. Archive and timestamp filtering
are supported by this CRD. It is compatible with Vertica server versions
24.2.0 and later. For instance, you may create the following CR to see
all the restore points for the backup archive in the database at
VerticaDB verticadb-sample:
```
apiVersion: vertica.com/v1beta1
kind: VerticaRestorePointsQuery
metadata:
  name: verticarestorepointsquery-sample
spec:
  verticaDBName: verticadb-sample
  filterOptions:
	archiveName: backup   
```
- The VerticaScrutinize CRD allows you to execute and gather data from
the scrutinize command against a live database. This eliminates the need
to run the scrutinize command in the Vertica server pods. The scrutinize
tarball is stored in a volume, and a pod remains active for a period of
time, allowing you to inspect and extract the tarball's contents. Note
that this feature is only available for Vertica servers running version
24.2.0 or later. For instance, create the following CR to run scrutinize
against the database specified in the verticadb-sample VerticaDB CR:
```
apiVersion: vertica.com/v1beta1
kind: VerticaScrutinize
metadata:
  name: verticascrutinize-sample
spec:
  verticaDBName: verticadb-sample
```

---------

Signed-off-by: GitHub <noreply@github.com>
Co-authored-by: jizhuoyu <44009608+jizhuoyu@users.noreply.github.com>
Co-authored-by: Roy Paulin <rnguetsopken@opentext.com>
Co-authored-by: chinhtranvan <47764275+chinhtranvan@users.noreply.github.com>
Co-authored-by: Zhuoyu Ji <zhuoyu.ji@microfocus.com>
Co-authored-by: chinhtranvan <tranc@cua.edu>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: spilchen <spilchen@users.noreply.github.com>
Co-authored-by: roypaulin <roy.paulin@microfocus.com>