Update push-vertica-server-image.sh for org change
The push-vertica-server-image.sh was updated to push/pull from the
opentext docker org. We also build the kafka-scheduler image in the
internal Jenkins pipeline, so we want to publish that image when a new
server is released. That image is a multi-arch image so we had to use
docker buildx imagetools to move the image to the public repository.