Mount scrutinize directory in server container (#733)
During the execution of the scrutinize process, it failed to collect
system table information due to the inability to create the staging
directory. This occurred because the directory is created from the HTTPS
service, which operates within the server container. However, the
container lacked access to the scrutinize mount path (/tmp/scrutinize).
This change addresses the problem by mounting the required directory in
the server container. Upon upgrading the operator with this fix, a
rolling update will be initiated across all statefulsets to incorporate
the new mount.

---------

Co-authored-by: Matt Spilchen <matt.spilchen@microfocus.com>