Change VerticaScrutinize to read password from file (#742)
Previously, we set the password secret environment variables to allow
scrutinize, when invoked via the VerticaScrutinize CRD, to read the
password from the secret. Now, when the password secret is in
Kubernetes, we mount it into the scrutinize init container and configure
the scrutinize command to read the password from a file. This eliminates
the need to create a role granting read access to the secret when the
secret is in Kubernetes.

---------

Co-authored-by: Matt Spilchen <mspilchen@opentext.com>