Update process check to exclude inactive PIDs (#745)
We have code to check if the Vertica process is running in the pod using
the 'ps' command. In some tests, even after deleting the pod in
admintools deployments, we noticed the Vertica process still running.
This isn't correct because usually, we have to restart it manually. I
suspect the Vertica PID might momentarily appear because it's terminated
and waiting to be cleared. I'm not completely sure. Therefore, I'm
updating the code to only find the PID for processes that are actively
running. If a PID is stopped or terminated, it won't be found. This
change will hopefully improve the pass rate for some upgrade tests.